requirements.txt
djangorestframework
stripe
dotenv